Move Compiler.hs one step up
This eliminates the old Compiler.hs that
is just a forwarding module.
